+++ app/com/crackers/repositories/EmailRepository.java
@@ -12,16 +12,16 @@ public interface EmailRepository extends GraphRepository<Email>
 {
 
     @Query("match (email)<-[:HAS_AN_EMAIL]-(user:User) where user.idUser = {idUser} and email.isPrimary = 1 and email.isDeleted = 0 return email")
-    Email getCreatedByMailId(@Param("idUser") Integer idUser);
+    Email getCreatedByMailId(@Param("idUser") Long idUser);
 
     @Query("match (email)<-[:HAS_AN_EMAIL]-(user:User) where user.idUser = {idUser} and email.isDeleted = 0 return email")
-    List<Email> getUsersMailId(@Param("idUser") Integer idUser);
+    List<Email> getUsersMailId(@Param("idUser") Long idUser);
 
     @Query("match (email)<-[:HAS_AN_EMAIL]-(user:User) where LOWER(email.email) like LOWER({email}) and user.idUserState = 1 and user.isDeleted = 0 and email.isDeleted = 0 return user.idUser")
-    Integer validateNewUserMailId(@Param("email") String email);
+    Long validateNewUserMailId(@Param("email") String email);
 
     @Query("match (email)<-[:HAS_AN_EMAIL]-(user:User) where user.idUserState = 1 and user.idUser = {idUser} and email.isPrimary = 1 and user.isDeleted = 0 and email.isDeleted = 0 return email.email")
-    String getEmailByIdUser(@Param("idUser") Integer idUser);
+    String getEmailByIdUser(@Param("idUser") Long idUser);
 
     @Query("match (emails:Email) where e.isDeleted = 0 return emails")
     List<Email> getEmails();
