+++ app/com/crackers/repositories/PasswordRepository.java
@@ -12,7 +12,7 @@ public interface PasswordRepository extends GraphRepository<Password>
 {
 
     @Query("match (previousPassword)<-[:HAS_PREVIOUS_PASSWORDS]-(user:User) where user.idUser = {idUser} and previousPassword.isDeleted = 0 and previousPassword.isExpired = 0 return previousPassword")
-    List<Password> getPasswordListById(@Param("idUser") Integer idUser);
+    List<Password> getPasswordListById(@Param("idUser") Long idUser);
 
     @Query("match (p:Password) where p.idPassword = {idPassword} and p.isDeleted = 0 and p.isExpired = 0 return p")
     Password getPasswordById(@Param("idPassword") Integer idPassword);
