+++ app/com/crackers/repositories/PhoneNumberRepository.java
@@ -12,7 +12,7 @@ public interface PhoneNumberRepository extends GraphRepository<PhoneNumber>
 {
 
     @Query("match (phoneNumbers)<-[:HAS_MULTIPLE_PHONE_NUMBERS]-(user:User) where user.idUser = {idUser} and phoneNumbers.isDeleted = 0 return phoneNumbers")
-    List<PhoneNumber> getUserPhoneNumbers(@Param("idUser") Integer idUser);
+    List<PhoneNumber> getUserPhoneNumbers(@Param("idUser") Long idUser);
 
     @Query("match (pn:PhoneNumber) where pn.isDeleted = 0 return pn")
     List<PhoneNumber> getPhoneNumbers();
