+++ app/com/crackers/repositories/UserContactDetailsRepository.java
@@ -12,7 +12,7 @@ public interface UserContactDetailsRepository extends GraphRepository<ContactDet
 {
 
     @Query("match (addresses)<-[:HAS_MULTIPLE_ADDRESSES]-(user:User) where user.idUser = {idUser} and addresses.isDeleted = 0 return addresses")
-    List<ContactDetails> getContactDetails(@Param("idUser") Integer idUser);
+    List<ContactDetails> getContactDetails(@Param("idUser") Long idUser);
 
     @Query("match (cd:ContactDetails) where cd.isDeleted = 0")
     List<ContactDetails> getAllContactDetails();
